backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.19
django-cacheops==7.0.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.21
django-filter==23.2
django-nine==0.2.7
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
elasticsearch==7.9.1
elasticsearch-dsl==7.4.1
Faker==18.9.0
frozenlist==1.4.1
funcy==1.18
gunicorn==20.1.0
idna==3.4
importlib_metadata==8.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
livekit-api==0.5.1
livekit-protocol==0.5.1
Markdown==3.4.3
MarkupSafe==2.1.2
multidict==6.0.5
packaging==23.1
Pillow==9.5.0
protobuf==5.27.2
mysqlclient
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
types-protobuf==4.25.0.20240417
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
yarl==1.9.4
zipp==3.19.2